Refactor Dockerfile for airflow_mysql container and update file name
This commit renames the file `docker_airflow_mysql.yaml` to `docker-compose.yaml` and makes changes to the services section of the file. The `otel-collector` service has been removed from the configuration.

Refactoring the Dockerfile and updating the file name improves the organization and clarity of the codebase.

Refactor code changes:
- Renamed `docker_airflow_mysql.yaml` to `docker-compose.yaml`
- Removed `otel-collector` service from the configuration